pyodbc/requirements.txt
pyodbc==5.1.0
python-dotenv==1.0.1
